pyproject.toml
[project]
name = "kerala-ayurveda-rag"
version = "0.1.0"
description = "Production-ready RAG system for Kerala Ayurveda with CRAG, Knowledge Graphs, and Multi-Agent Architecture"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Kerala Ayurveda RAG Team" }]
keywords = ["RAG", "Ayurveda", "LLM", "CRAG", "Knowledge Graph", "Multi-Agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core LLM & Embeddings
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    # Vector Store & Retrieval
    "chromadb>=0.4.22",
    "rank-bm25>=0.2.2",
    # LangChain Ecosystem
    "langchain>=0.1.11",
    "langchain-openai>=0.0.8",
    "langchain-community>=0.0.24",
    "langgraph>=0.0.26",
    # Evaluation
    "ragas>=0.1.4",
    # Data Processing
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "tiktoken>=0.6.0",
    # Knowledge Graph
    "networkx>=3.2.1",
    # API & UI
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "gradio>=4.16.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.2",
    "tenacity>=8.2.3",
    "tqdm>=4.66.0",
    "langchain-google-genai>=4.1.1",
    "google-generativeai>=0.8.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "ipython>=8.20.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "ipython>=8.20.0",
    "jupyter>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.uv
  | build
  | dist
  | __pycache__
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # do not perform function calls in argument defaults
    "B904", # raise from
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
